Isolated the go doc examples in a separate file
