+++ osmshortlink_test.go
@@ -1,7 +1,6 @@
 package osmshortlink
 
 import (
-	"fmt"
 	"math"
 	"testing"
 )
@@ -199,31 +198,4 @@ func TestDecodeShortLinkString(t *testing.T) {
 			}
 		})
 	}
-}
-
-func ExampleCreate() {
-	shortLink, err := Create(46.05141, 14.50604, 17)
-	if err != nil {
-		panic(err)
-	}
-	fmt.Println(shortLink)
-	// Output: https://osm.org/go/0Ik3VNr_A-?m
-}
-
-func ExampleEncode() {
-	shortLink, err := Encode(46.05141, 14.50604, 17)
-	if err != nil {
-		panic(err)
-	}
-	fmt.Println(shortLink)
-	// Output: 0Ik3VNr_A-
-}
-
-func ExampleDecode() {
-	latitude, longitude, zoom, err := Decode("0Ik3VNr_A-")
-	if err != nil {
-		panic(err)
-	}
-	fmt.Println(latitude, longitude, zoom)
-	// Output: 46.05140447616577 14.506051540374756 17
 }
